# docs/swagger.yaml
definitions:
...
  models.Profile:
    properties:
      description:
        type: string
      id:
        type: string
      img:
        type: string
      login:
        type: string
      phone:
        type: string
    type: object
...
  models.SignInPayload:
    properties:
      login:
        maxLength: 30
        minLength: 6
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
    required:
    - login
    - password
    type: object
  models.SignUpPayload:
    properties:
      login:
        maxLength: 30
        minLength: 6
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - login
    - password
    - phone
    type: object
  models.UpdateProfileDataPayload:
    properties:
      passwords:
        properties:
          newPass:
            maxLength: 32
            minLength: 8
            type: string
          oldPass:
            maxLength: 32
            minLength: 8
            type: string
        type: object
      phone:
        type: string
    type: object
...
  responser.response:
    properties:
      error: {}
      status:
        type: string
    type: object
info:
  contact:
    name: Dima
    url: http://t.me/belozerovmsk
  description: API server for ZuZu.
  title: ZuZu Backend API
paths:
...
  /api/auth/check_auth:
    get:
      consumes:
      - application/json
      description: Check user is logged in
      parameters:
      - description: X-CSRF-Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      - description: Cookie
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile
          schema:
            $ref: '#/definitions/models.Profile'
        "401":
          description: Unauthorized
        "429":
          description: Too Many Requests
      security:
      - AuthKey: []
      summary: CheckAuth
      tags:
      - Auth
  /api/auth/logout:
    get:
      consumes:
      - application/json
      description: Logout from Account
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      summary: Logout
      tags:
      - Auth
  /api/auth/signin:
    get:
      description: Login to Account
      produces:
      - application/json
      responses:
        "200":
          description: Profile
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: error messege
          schema:
            $ref: '#/definitions/responser.response'
        "429":
          description: Too Many Requests
      summary: SignIn(get)
      tags:
      - Auth
    post:
      consumes:
      - application/json
      description: Login to Account
      parameters:
      - description: SignInPayload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SignInPayload'
      - description: X-CSRF-Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      - description: Cookie
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: error messege
          schema:
            $ref: '#/definitions/responser.response'
        "429":
          description: Too Many Requests
      summary: SignIn
      tags:
      - Auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Create Account
      parameters:
      - description: SignUpPayload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SignUpPayload'
      - description: X-CSRF-Token
        in: header
        name: X-CSRF-Token
        required: true
        type: string
      - description: Cookie
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: error messege
          schema:
            $ref: '#/definitions/responser.response'
        "429":
          description: Too Many Requests
      summary: SignUp
      tags:
      - Auth
...
  /api/profile/{id}:
    get:
      consumes:
      - application/json
      description: Get profile by ID
      parameters:
      - description: Profile UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: error messege
          schema:
            $ref: '#/definitions/responser.response'
        "429":
          description: Too Many Requests
      summary: GetProfile
      tags:
      - Profile
  /api/profile/update-data:
    post:
      consumes:
      - application/json
      description: Update profile data
      parameters:
      - description: UpdateProfileDataPayload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileDataPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Profile
          schema:
            $ref: '#/definitions/models.Profile'
        "400":
          description: error messege
          schema:
            $ref: '#/definitions/responser.response'
        "401":
          description: Unauthorized
        "429":
          description: Too Many Requests
      summary: UpdateProfileData
      tags:
      - Profile
  /api/profile/update-photo:
    post:
      consumes:
      - application/json
      description: Update profile photo
      parameters:
      - description: photo
        in: body
        name: input
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Profile
          schema:
            $ref: '#/definitions/models.Profile'
        "401":
          description: Unauthorized
        "413":
          description: Request Entity Too Large
        "429":
          description: Too Many Requests
      summary: UpdatePhoto
      tags:
      - Profile
...
swagger: "2.0"